@@ -7,25 +7,24 @@
     "@emotion/styled": "^11.11.0",
     "@iconify/icons-mdi": "^1.2.46",
     "@iconify/react": "^4.1.0",
-    "@material-ui/core": "^4.12.4",
     "@microsoft/fetch-event-source": "^2.0.1",
     "@mui/icons-material": "^5.11.16",
-    "@mui/material": "^5.13.5",
+    "@mui/material": "^5.15.3",
     "@react-google-maps/api": "^2.19.2",
     "axios": "^1.4.0",
     "bootstrap": "^5.3.0",
     "meyda": "^4.3.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.8.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-google-places-autocomplete": "^4.0.1",
     "react-hook-form": "^7.49.0",
     "react-phone-number-input": "^3.3.7",
-    "react-resizable": "^3.0.5",
     "react-router-dom": "^6.13.0",
     "react-scripts": "^5.0.1",
     "react-spring-bottom-sheet": "^3.4.1",
     "react-stomp": "^5.1.0",
+    "react-table-column-resizer": "^1.2.3",
     "react-toastify": "^9.1.3",
     "react-widgets": "^5.8.4",
     "sass": "^1.62.1",
